requirements.txt
requests~=2.32.3
DrissionPage~=4.1.1.2
fonttools~=4.59.0
protobuf~=5.29.2
pillow~=11.1.0
ddddocr~=1.5.6
curl_cffi~=0.13.0